Apply pdfium_noshorten_config to xfa/fwl
Fall back to checked_cast<> for some ranges where -1 may be
required as a sentinel. Convert other indices to size_t where
possible.

-- tidy stringview construction while at it.

Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/90710
Commit-Queue: Tom Sepez <tsepez@chromium.org>

+++ xfa/fwl/cfwl_datetimepicker.cpp
@@ -158,7 +158,7 @@ WideString CFWL_DateTimePicker::GetEditText() const {
   return m_pEdit ? m_pEdit->GetText() : WideString();
 }
 
-int32_t CFWL_DateTimePicker::GetEditTextLength() const {
+size_t CFWL_DateTimePicker::GetEditTextLength() const {
   return m_pEdit ? m_pEdit->GetTextLength() : 0;
 }
